Move CUDA to an extension (#236)
* move CUDA to an extension

* fix CUDA deps

* fix exports

* no Requires

* reformat toml

* Ext

* canonical

* generalise copy_to_output

* add the actual extension

* fix
