+++ Project.toml
@@ -20,15 +20,21 @@ OffsetArrays = "6fe1bfb0-de20-5000-8ca7-80f57d26f881"
 REPL = "3fa0cd96-eef1-5676-8a61-b3b8758bbffb"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
-Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 Setfield = "efcf1570-3423-57d1-acb7-fd33fddbac46"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 Stencils = "264155e8-78a8-466a-aa59-c9b28c34d21a"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 UnicodeGraphics = "ebadf6b4-db70-5817-83da-4a19ad584e34"
 
+[weakdeps]
+CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
+
+[extensions]
+DynamicGridsCUDAExt = "CUDA"
+
 [compat]
 Adapt = "2, 3"
+CUDA = "5"
 Colors = "0.9, 0.10, 0.11, 0.12"
 ConstructionBase = "1"
 Crayons = "4"
@@ -40,7 +46,6 @@ KernelAbstractions = "0.9"
 ModelParameters = "0.3, 0.4"
 OffsetArrays = "0.11, 1"
 Reexport = "0.2, 1"
-Requires = "1"
 Setfield = "0.6, 0.7, 0.8, 1"
 StaticArrays = "0.12, 1"
 Stencils = "0.2"
