+++ src/DynamicGrids.jl
@@ -11,7 +11,6 @@ import Adapt,
       ConstructionBase,
       FreeTypeAbstraction,
       Reexport,
-      Requires,
       REPL
 
 using Colors,
@@ -64,7 +63,7 @@ export AbstractRuleset, Ruleset, StaticRuleset
 
 export AbstractSimData
 
-export Processor, SingleCPU, ThreadedCPU, CPUGPU
+export Processor, SingleCPU, ThreadedCPU, CPUGPU, CuGPU
 
 export PerformanceOpt, NoOpt, SparseOpt
 
@@ -146,8 +145,6 @@ include("show.jl")
 function __init__()
     global terminal
     terminal = REPL.Terminals.TTYTerminal(get(ENV, "TERM", Base.Sys.iswindows() ? "" : "dumb"), stdin, stdout, stderr)
-
-    Requires.@require CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba" include("cuda.jl")
 end
 
 end
